viewId converted to enum, arythmetic comparisions replaced by bool func
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7461 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ renderer/tr_lightrun.cpp
@@ -140,7 +140,7 @@ void R_ModulateLights_f( const idCmdArgs &args ) {
 	}
 
 	int count = 0;
-	float modulate[3] = {atof(args.Argv(1)), atof(args.Argv(2)), atof(args.Argv(3))};
+	double modulate[3] = {atof(args.Argv(1)), atof(args.Argv(2)), atof(args.Argv(3))};
 	idRenderLightLocal *light;
 
 	for (int i = 0 ; i < tr.primaryWorld->lightDefs.Num() ; i++ ) {
